feat: show now upcoming movies on list

+++ app/src/main/java/com/duhan/movieapp/feature_movie/data/data_source/network/model/UpcomingResult.kt
@@ -15,6 +15,6 @@ data class UpcomingResult(
     @SerializedName("release_date") var releaseDate: String? = null,
     @SerializedName("title") var title: String? = null,
     @SerializedName("video") var video: Boolean? = null,
-    @SerializedName("vote_average") var voteAverage: Int? = null,
+    @SerializedName("vote_average") var voteAverage: Double? = null,
     @SerializedName("vote_count") var voteCount: Int? = null
 )
